finish load discounts for static user, wait for merging to convert to dynamic

+++ hair_booking/app/src/main/java/com/example/hair_booking/services/db/DbDiscountServices.kt
@@ -37,7 +37,7 @@ class DbDiscountServices(private var dbInstance: FirebaseFirestore?): DatabaseAb
                     document.data?.get("description") as String,
                     document.data?.get("dateApplied") as String,
                     document.data?.get("dateExpired") as String,
-                    document.data?.get("percent") as Long
+                    document.data?.get("percent") as Double
                 )
 
                 discountList.add(discount)
